refactor: use a single lz conf for all networks
